Add write buffer to improve performance
